WIP: introduce new module genesis_config_presets in the runtime
