+++ node/service/src/chain_spec/vara.rs
@@ -139,80 +139,3 @@ pub fn local_testnet_config() -> Result<ChainSpec, String> {
         .with_properties(vara_dev_properties())
         .build())
 }
-
-/// Configure initial storage state for FRAME modules.
-fn testnet_genesis(
-    initial_authorities: Vec<(
-        AccountId,
-        AccountId,
-        BabeId,
-        GrandpaId,
-        ImOnlineId,
-        AuthorityDiscoveryId,
-    )>,
-    root_key: AccountId,
-    endowed_accounts: Vec<AccountId>,
-    bank_account: AccountId,
-    _enable_println: bool,
-) -> serde_json::Value {
-    const ENDOWMENT: u128 = 1_000_000 * TOKEN;
-    const STASH: u128 = 100 * TOKEN;
-    const MIN_NOMINATOR_BOND: u128 = 50 * TOKEN;
-
-    let _num_endowed_accounts = endowed_accounts.len();
-
-    let mut balances = endowed_accounts
-        .iter()
-        .map(|k: &AccountId| (k.clone(), ENDOWMENT))
-        .chain(initial_authorities.iter().map(|x| (x.0.clone(), STASH)))
-        .collect::<Vec<_>>();
-
-    balances.push((bank_account, EXISTENTIAL_DEPOSIT));
-
-    serde_json::json!({
-        "balances": {
-            "balances": balances,
-        },
-
-        "session": {
-            "keys": initial_authorities
-                    .iter()
-                    .map(|x| {
-                        (
-                            x.0.clone(),
-                            x.0.clone(),
-                            session_keys(x.2.clone(), x.3.clone(), x.4.clone(), x.5.clone()),
-                        )
-                    })
-                    .collect::<Vec<_>>(),
-        },
-        "staking": {
-            "validatorCount": initial_authorities.len() as u32,
-            "minimumValidatorCount": 4,
-            "stakers": initial_authorities
-            .iter()
-            .map(|x| (x.0.clone(), x.1.clone(), STASH, StakerStatus::<AccountId>::Validator))
-            .collect::<Vec<_>>(),
-            "invulnerables": initial_authorities.iter().map(|x| x.0.clone()).collect::<Vec<_>>(),
-            "forceEra": Forcing::ForceNone,
-            "slashRewardFraction": Perbill::from_percent(10),
-            "minNominatorBond": MIN_NOMINATOR_BOND,
-        },
-        "nominationPools": {
-            "minCreateBond": 10 * ECONOMIC_UNITS,
-            "minJoinBond": ECONOMIC_UNITS,
-        },
-        "stakingRewards": {
-            "nonStakeable": Perquintill::from_rational(4108_u64, 10_000_u64), // 41.08%
-            "idealStake": Perquintill::from_percent(85), // 85%
-            "targetInflation": Perquintill::from_rational(578_u64, 10_000_u64), // 5.78%
-        },
-        "babe": {
-            "epochConfig": Some(vara_runtime::BABE_GENESIS_EPOCH_CONFIG),
-        },
-        "sudo": {
-            // Assign network admin rights.
-            "key": Some(root_key),
-        },
-    })
-}
