+++ runtime/vara/Cargo.toml
@@ -19,6 +19,7 @@ const-str.workspace = true
 log.workspace = true
 parity-scale-codec.workspace = true
 scale-info = { workspace = true, features = ["derive"] }
+serde_json.workspace = true
 
 # Frame deps
 frame-support.workspace = true
@@ -192,6 +193,7 @@ std = [
 	"gear-runtime-common/std",
 	"runtime-primitives/std",
 	"scale-info/std",
+	"serde_json/std",
 	"sp-api/std",
 	"sp-arithmetic/std",
 	"sp-authority-discovery/std",
